Crate io_at

Modules

os
OS specific implimentations of WriteAt and/or ReadAt

Structs

At
Use an offset as the base for all ReadAt and WriteAt accesses. Where Take modifies then end, At modifies the start.

BlockLimitWrite
Limit the amount of data accepted in a single call to WriteAt (ReadAt is unaffected)

Cursor
Adapt a WriteAt and/or ReadAt to a Write and/or Read by tracking a single internal offset and modifying it.

LockedSeek
Allow a Seek + (Read and/or Write) to impliment WriteAt and/or ReadAt

Take
Limit the maximum offset of a WriteAt and/or ReadAt.

Traits

ReadAt
Read data at an offset

WriteAt
Write data at an offset

Type Definitions

Result
A specialized Result type for I/O operations.