Trait io_at::os::AsRaw

pub trait AsRaw {
    fn as_raw_fd(&self) -> i32;
}

A trait to extract the raw unix file descriptor from an underlying object.

This is only available on unix platforms and must be imported in order to call the method. Windows platforms have a corresponding AsRawHandle and AsRawSocket set of traits.

Required methods

fn as_raw_fd(&self) -> i32

Extracts the raw file descriptor.

This method does not pass ownership of the raw file descriptor to the caller. The descriptor is only guaranteed to be valid while the original object has not yet been destroyed.

Example

use std::fs::File;
use std::os::unix::io::{AsRawFd, RawFd};

let mut f = File::open("foo.txt")?;
// Note that `raw_fd` is only valid as long as `f` exists.
let raw_fd: RawFd = f.as_raw_fd();

Loading content...