Struct io_at::LockedSeek

pub struct LockedSeek<T: Seek> {
    // some fields omitted
}

Allow a Seek + (Read and/or Write) to impliment WriteAt and/or ReadAt

While in most cases more specific adaptations will be more efficient, this allows any Seek to impliment WriteAt or ReadAt with not additional restrictions.